refactor(api_demo): refactored old api_example.py logic to separate modules within /api folder
