+++ api/dependencies.py
@@ -1,0 +1,24 @@
+from fastapi import WebSocket
+from .utils.phrases import get_random_phrase
+import asyncio
+import json
+
+class ConnectionManager:
+    active_connections: list[WebSocket] = []
+
+    async def connect(self, websocket: WebSocket):
+        await websocket.accept()
+        self.active_connections.append(websocket)
+
+    def disconnect(self, websocket: WebSocket):
+        self.active_connections.remove(websocket)
+
+    async def send_personal_message(self, message: dict, websocket: WebSocket):
+        message_json = json.dumps(message)
+        await websocket.send_text(message_json)
+
+    async def broadcast(self, message: str):
+        for connection in self.active_connections:
+            await connection.send_text(message)
+
+manager = ConnectionManager()
