+++ api/models/query.py
@@ -1,0 +1,4 @@
+from pydantic import BaseModel
+
+class QueryParams(BaseModel):
+    query: str
